# WMS.Angular/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WMS API v1",
    "version": "v1"
  },
  "paths": {
...
    "/rack": {
      "get": {
        "tags": [
          "Racks"
        ],
        "summary": "Get racks list",
        "description": "Retrieves all racks with optional filtering by warehouse.",
        "operationId": "GetRacks",
        "parameters": [
          {
            "name": "warehouseId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RackSummaryDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Racks"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRackDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/rack/warehouse/{id}": {
      "get": {
        "tags": [
          "Racks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RackSummaryDto"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/rack/{id}": {
      "get": {
        "tags": [
          "Racks"
        ],
        "operationId": "GetRack",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Racks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRackDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Racks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/rack/bin/{id}": {
      "get": {
        "tags": [
          "Racks"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/warehouse/{id}": {
      "get": {
        "tags": [
          "WMS.Api"
        ],
        "operationId": "GetWarehouse",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "WMS.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWarehouseDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "WMS.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/bin/rack/{id}": {
      "get": {
        "tags": [
          "WMS.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BinSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/bin/v2/rack/{id}": {
      "get": {
        "tags": [
          "WMS.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BinSummaryDtoPaginatedResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/bin/{id}": {
      "get": {
        "tags": [
          "WMS.Api"
        ],
        "summary": "Get bin details by ID",
        "description": "Retrieves the structural configuration and metadata of a specific bin location using its primary key ID.",
        "operationId": "GetBin",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BinDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "WMS.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBinDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "WMS.Api"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BinDetailsDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "rackId": {
            "type": "integer",
            "format": "int32"
          },
          "bayId": {
            "type": "integer",
            "format": "int32"
          },
          "levelId": {
            "type": "integer",
            "format": "int32"
          },
          "binId": {
            "type": "integer",
            "format": "int32"
          },
          "binHashCode": {
            "type": "integer",
            "format": "int32"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "BinSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "warehouse": {
            "type": "string",
            "nullable": true
          },
          "rack": {
            "type": "string",
            "nullable": true
          },
          "bay": {
            "type": "string",
            "nullable": true
          },
          "level": {
            "type": "string",
            "nullable": true
          },
          "binName": {
            "type": "string",
            "nullable": true
          },
          "binHashCode": {
            "type": "string",
            "nullable": true
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "BinSummaryDtoPaginatedResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BinSummaryDto"
            },
            "nullable": true
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "CreateRackDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "warehouseId": {
            "type": "integer",
            "format": "int32"
          },
          "bayId": {
            "type": "integer",
            "format": "int32"
          },
          "levelId": {
            "type": "integer",
            "format": "int32"
          },
          "rackFloorMapObject": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RackFloorMapObjectDetails"
            },
            "nullable": true
          },
          "isFloorDocking": {
            "type": "boolean"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "RackFloorMapObjectDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "rackId": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "left": {
            "type": "number",
            "format": "float"
          },
          "top": {
            "type": "number",
            "format": "float"
          },
          "width": {
            "type": "number",
            "format": "float"
          },
          "height": {
            "type": "number",
            "format": "float"
          },
          "text": {
            "type": "string",
            "nullable": true
          },
          "fill": {
            "type": "string",
            "nullable": true
          },
          "stroke": {
            "type": "string",
            "nullable": true
          },
          "strokeWidth": {
            "type": "number",
            "format": "double"
          },
          "fontSize": {
            "type": "integer",
            "format": "int32"
          },
          "scaleX": {
            "type": "number",
            "format": "float"
          },
          "scaleY": {
            "type": "number",
            "format": "float"
          },
          "angle": {
            "type": "number",
            "format": "float"
          }
        },
        "additionalProperties": false
      },
      "RackSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "warehouse": {
            "type": "string",
            "nullable": true
          },
          "bay": {
            "type": "string",
            "nullable": true
          },
          "level": {
            "type": "string",
            "nullable": true
          },
          "rackFloorMapObject": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RackFloorMapObjectDetails"
            },
            "nullable": true
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateBinDto": {
        "type": "object",
        "properties": {
          "rackId": {
            "type": "integer",
            "format": "int32"
          },
          "bayId": {
            "type": "integer",
            "format": "int32"
          },
          "levelId": {
            "type": "integer",
            "format": "int32"
          },
          "binId": {
            "type": "integer",
            "format": "int32"
          },
          "binHashCode": {
            "type": "integer",
            "format": "int32"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "UpdateRackDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "warehouseId": {
            "type": "integer",
            "format": "int32"
          },
          "bayId": {
            "type": "integer",
            "format": "int32"
          },
          "levelId": {
            "type": "integer",
            "format": "int32"
          },
          "rackFloorMapObject": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RackFloorMapObjectDetails"
            },
            "nullable": true
          },
          "isFloorDocking": {
            "type": "boolean"
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UpdateWarehouseDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "floorPlanImage": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "dateAdded": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "description": "ASP.NET Core Identity Application Cookie",
        "name": ".Skybest.WMS.Auth",
        "in": "cookie"
      }
    }
  },
  "security": [
    {
      "CookieAuth": []
    }
  ]
}